Add resources files
- add main color
- add rocket, logo images
- add sound files
- add splash configurations
